+++ jubatus/core/clustering/eigen_feature_mapper.hpp
@@ -50,16 +50,16 @@ class eigen_feature_mapper {
   jubatus::util::data::unordered_map<size_t, std::string> rmap_;
 
   void insert(
-      const std::pair<std::string, float>& item,
+      const std::pair<std::string, double>& item,
       bool update_map,
       eigen_svec_t& dst);
 
   void insertc(
-      const std::pair<std::string, float>&,
+      const std::pair<std::string, double>&,
       eigen_svec_t& dst) const;
 
   void rinsert(
-      const std::pair<int, float>&,
+      const std::pair<int, double>&,
       common::sfv_t& dst) const;
 
   int d_;
